表结构更新

+++ src/main/java/com/cj/lottery/domain/view/LotteryActivityInfoVo.java
@@ -29,5 +29,5 @@ public class LotteryActivityInfoVo {
     private int consumerNum;
 
     @ApiModelProperty("活动标识:0:普通活动 1:新人活动")
-    private Byte activityFlag;
+    private Integer activityFlag;
 }
